requirements.txt
Flask>=3.0
gunicorn>=21.2
python-dotenv>=1.0
crewai>=0.40
pydantic>=1.10
web3>=6.0
pandas>=2.1
openpyxl>=3.1
boto3>=1.28